pyproject.toml
[project]
name = "openai-dummy-api"
version = "0.1.0"
description = "OpenAI-compatible API proxy for capturing prompts and tools"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "tiktoken>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
